pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[project]
name = "autonomous-coding-agent"
version = "0.1.0"
description = "Deterministic autonomous delivery harness with bounded AI strategy."
requires-python = ">=3.11"
dependencies = [
  "jsonschema>=4.0.0",
  "PyYAML>=6.0.0",
]

[tool.setuptools.packages.find]
include = ["supervisor*"]
